requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
deprecation==2.1.0
idna==3.4
img2pdf==0.4.4
lxml==4.9.2
packaging==23.0
pikepdf==6.2.8.post1
Pillow==9.4.0
requests==2.28.2
urllib3==1.26.14
